Separate the logic in the drawable lyric list.
